requirements.txt
requests
beautifulsoup4
python-dotenv
lxml
python-docx
pandas
openpyxl
python-Levenshtein
pymupdf
pdfplumber
pytest
